% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zero_obstime.R
\name{add_zero_obstime}
\alias{add_zero_obstime}
\title{Add row with 0 observation time}
\usage{
add_zero_obstime(obs_time, focal_animal)
}
\arguments{
\item{obs_time}{data frame created by obs_time_calculator()}

\item{focal_animal}{character string of focal ID}
}
\value{
A data frame with an extra row
}
\description{
function to add a row to obs_time for dra and fro who had 0 observation hours
(were not focalled)
}
