% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_part.R
\name{check_part}
\alias{check_part}
\title{Check partner}
\usage{
check_part(x)
}
\arguments{
\item{x}{A data frame. Needs columns "focal_animal" and "action_partner"}
}
\value{
Adds a new column "check_partner" to x, with "yes" or "no" if action partner was also a focal that day.
Also adds "rowid" column. Need this so that there are no duplicate rows, so that merging later works properly
}
\description{
Check if action_partner was also a focal animal on the date of the action
}
