% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decider.R
\name{decider}
\alias{decider}
\title{Decider}
\usage{
decider(x)
}
\arguments{
\item{x}{A data frame of interactions. Typically generated by the "duration decider" function. Needs columns "focal_dyad_count" and "partner_dyad_count". (created by get_p_dyad_count).}
}
\value{
Adds new "double_interaction" column with decision on wether to "include" or "exclude" a particular interaction event.
}
\description{
A helper function. Highlights which double interactions to keep or exclude. Used inside "get_int_decision".
}
