% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decider_diff_count.R
\name{decider_diff_count}
\alias{decider_diff_count}
\title{Decider different count}
\usage{
decider_diff_count(focal_dyad_count, partner_dyad_count)
}
\arguments{
\item{focal_animal}{ID of animal that was being focalled, as a character string}

\item{action_partner}{ID of action partner of interaction, as a character string}
}
\value{
A string with the decision, either "include" or "exclude"
}
\description{
Helper function for deciding which focal interaction to keep, when dyad counts are different. Will decide to keep whichever focal has the highest dyad count.
}
\details{
This function will be used inside the "decider" function.
}
