% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decider_same_count.R
\name{decider_same_count}
\alias{decider_same_count}
\title{Decider same count}
\usage{
decider_same_count(focal_animal, action_partner)
}
\arguments{
\item{focal_animal}{ID of animal that was being focalled, as a character string}

\item{action_partner}{ID of action partner of interaction, as a character string}
}
\value{
A string with the decision, either "include" or "exclude"
}
\description{
Helper Function for deciding which focal interaction to keep, when dyad counts are the same. Will decide to keep if its ID comes before the action partner alphabetically
}
\details{
#' This function will be used inside the "decider" function.
}
