% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_sequence_NF.R
\name{elo_sequence_NF}
\alias{elo_sequence_NF}
\title{Modified elo.seq function}
\usage{
elo_sequence_NF(datafile = "c:\\\\example data.csv", sep = ",",
  startingvalue = 1000, constant_k = 200, priorElo = list(),
  priorRanks = list(), priorRankCategory = list(), priorRankIndex = 0,
  outcome = 1)
}
\arguments{
\item{datafile}{the path and file name of the csv data file with the interaction data; alternatively, will read existing dataframe}

\item{sep}{separator use, if datafile is a .csv file.}

\item{startingvalue}{the amount of points each individual has at the beginning of the rating process (default is 1000, see example in Albers and de Vries, 2001)}

\item{constant_k}{specify the amount your constant 'k' should take (default is 200, see example in Albers and de Vries, 2001)}

\item{priorElo}{a list of IDs (case-sensitive), and associated initial Elo-rating values (e.g. ID1 = 2200, ID2 = 1500).}

\item{priorRanks}{a list of IDs (case-sensitive), and associated starting ordinal ranks (e.g. ID1 = 1, ID2 = 2). Processed only if 'priorElo' not specified}

\item{priorRankCategory}{a list of IDs (case sensitive), and associate rank categories ("alpha", high", "medium", "low"): will accept initial letters & is case insensitive (e.g. "a", "H" ...). Individuals not listed will start with default Elo-rating (i.e. 'startingvalue'). Processed only if 'priorElo' & 'priorRanks' not specified.}

\item{priorRankIndex}{default is 0 for a distance between starting ranks equal to ‘k’; for a degree of despotism, such that elo-rating distances between high ranking males and others are exaggerated, enter reciprocal power: 0.3 works well.}

\item{outcome}{1 = first individual wins; 2 = second individual wins; 0 = draw}
}
\description{
elo.seq funtion modified by Newton-Fisher, 2017. Original function from Neuman et al., 2011.
}
