% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excl_dbl_int.R
\name{excl_dbl_int}
\alias{excl_dbl_int}
\title{Exclude double interactions}
\usage{
excl_dbl_int(df, join_by)
}
\arguments{
\item{df}{A data frame of interactions. Typically generated by the "duration decider" function. Needs columns "focal_dyad_count" and "partner_dyad_count". (created by get_p_dyad_count).}
}
\value{
A dataframe with double interactions filtered out.
}
\description{
Function to exclude double focal and action_partner interactions used for ap, bc, and gr, from duration calculator i.e. when they were focalled on the same day, and interact, keep only one
}
