% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_windows.R
\name{extract_windows}
\alias{extract_windows}
\title{Function to extract sample windows}
\usage{
extract_windows(df, ur, window_start, window_end)
}
\arguments{
\item{df}{Data frame of focal observations}

\item{ur}{Vector of urine sample numbers}

\item{window_start}{Start of window relative to time of urine collection (in seconds)}

\item{window_end}{End of window relative to time of urine collection (in seconds)}
}
\value{
Retruns a data frame of all sample windows.
}
\description{
For each sample, a subset data frame is created with all behaviours that 
occurred between window_start and window_end. Adds a column called "sample_ref",
which identifies which sample the window belongs to. Then stacks all sample 
windows on top of each other into one big data frame. 

Currently needs these columns with exact names: "action_time_of_day_sec", 
"focal_animal", "date"
}
