% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_day_elo.R
\name{get_day_elo}
\alias{get_day_elo}
\title{Get Elo at the end of day}
\usage{
get_day_elo(elo_sequence, ids)
}
\arguments{
\item{elo_sequence}{elo sequence generated by elo_model3()}

\item{ids}{vector of unique individual IDs}
}
\value{
A data frame with elo-ratings per day per ID
}
\description{
function to get elo-rating at the end of the day per individual
}
\examples{
elo.ama <- get_day_elo(ama_seq3.2, ama_c)

}
