% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_dyad_count.R
\name{get_p_dyad_count}
\alias{get_p_dyad_count}
\title{Get partner dyad count}
\usage{
get_p_dyad_count(x)
}
\arguments{
\item{x}{A data frame. Needs columns "date", "focal_animal" and "action_partner"}
}
\value{
Adds new column "partner_dyad_count" to x. It is 0 when focal and partner were not focalled on the same day, and >0 when they were. The number indicates how many times they interacted that day from the focal animal's perspective.
}
\description{
Get partner dyad count
}
