% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_row.R
\name{insert_row}
\alias{insert_row}
\title{Insert a new row into a data frame}
\usage{
insert_row(df, newrow, r)
}
\arguments{
\item{df}{a data frame}

\item{newrow}{new row to insert, must have same columns as df}

\item{r}{row index indicating where new row will be inserted}
}
\value{
Returns a data frame with an extra row inserted at specified index
}
\description{
Insert a new row into a data frame
}
\examples{
# example dataframe
d <- as.data.frame(matrix(seq(20), nrow=5, ncol=4))

# row number that you want to insert
r <- 3

# create a new row to insert
newrow <- seq(4)
# can be a subset of original dataframe
newrow2 <- d[1, ]

# insert row
insert_row(d, newrow, r)
insert_row(d, newrow2, r)
}
