% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cor.R
\name{plot_cor}
\alias{plot_cor}
\title{Plot correlation between two vars}
\usage{
plot_cor(dataset, xvar, yvar)
}
\arguments{
\item{dataset}{a data frame or tibble}

\item{xvar}{variable to plot on the x axis, as a character string. NA values are silently dropped.}

\item{yvar}{variable to plot on the y axis, as a character string. NA values are silently dropped.}
}
\value{
A ggplot object which may be further customised with e.g. theme() options.
}
\description{
Plots a scatter plot of x and y and includes a regression line.
}
\examples{
plot_cor(mtcars, "wt", "mpg")
}
