% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ln_grp.R
\name{plot_ln_grp}
\alias{plot_ln_grp}
\title{Plot line graph by group}
\usage{
plot_ln_grp(dataframe, xvar, yvar, grp_var = "focal_animal")
}
\arguments{
\item{dataframe}{a data frame or tibble}

\item{xvar}{variable to plot on the x axis, as a character string}

\item{yvar}{variable to plot on the y axis, as a character string. NA values are silently dropped.}

\item{grp_var}{variable to group by, as a character string. Default set to "focal_animal".}
}
\value{
A ggplot object which may be further customised with e.g. theme() options.
}
\description{
Plots a line graph connecting mean e.g. hormone values per day/week/month per subject. Manullay scales point shape and colour of up to 14 subjects so that each is easily identifiable.
}
