% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_to_hms.R
\name{sec_to_hms}
\alias{sec_to_hms}
\title{Function that converts a number of seconds to a character format hh:m:ss}
\usage{
sec_to_hms(secs)
}
\arguments{
\item{secs}{a number of seconds to be formatted}
}
\description{
This function converts a number of seconds to a character format hh:m:ss
}
