% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_int_decision.R
\name{get_int_decision}
\alias{get_int_decision}
\title{Get decision}
\usage{
get_int_decision(x)
}
\arguments{
\item{x}{A data frame of interactions. Typically generated by the "duration decider" function. Needs columns "focal_dyad_count" and "partner_dyad_count". (created by get_p_dyad_count).}
}
\description{
Highlights which double interactions to keep or exclude.
}
